🦐 feat(task-viewer): Complete React rewrite with modern tabbed interface
## 🎨 Major UI Overhaul
- **Modern Tab Interface**: Professional browser-style tabs with drag & drop reordering
- **Connected Design**: Seamless visual connection between tabs and content
- **Improved Layout**: Search and controls repositioned for better workflow

## ⚡ Enhanced Functionality
- **Configurable Auto-refresh**: Choose intervals from 5 seconds to 5 minutes
- **Advanced Search**: Real-time filtering across all task fields
- **Sortable Columns**: Click headers to sort by any column
- **Hot Reload Development**: Instant updates during development

## 🔧 Technical Improvements
- **React Architecture**: Complete rewrite using React 19 + Vite
- **TanStack Table**: Professional table component with pagination
- **Responsive Design**: Mobile-first approach with breakpoint optimization
- **Performance**: Optimized rendering and efficient state management

## 🎯 UX Enhancements
- **Visual Feedback**: Hover states, loading indicators, and transitions
- **Accessibility**: Full keyboard navigation and screen reader support
- **Professional Styling**: Cohesive design language throughout
- **Intuitive Controls**: Logical grouping and placement of interface elements
