+++ tools/task-viewer/src/components/TaskTable.jsx
@@ -1,0 +1,212 @@
+import React, { useMemo } from 'react';
+import {
+  useReactTable,
+  getCoreRowModel,
+  getFilteredRowModel,
+  getPaginationRowModel,
+  getSortedRowModel,
+  flexRender,
+} from '@tanstack/react-table';
+
+function TaskTable({ data, globalFilter, onGlobalFilterChange }) {
+  const columns = useMemo(() => [
+    {
+      accessorKey: 'taskNumber',
+      header: '#',
+      cell: ({ row }) => (
+        <span className="task-number">
+          TASK {row.index + 1}
+        </span>
+      ),
+      size: 80,
+    },
+    {
+      accessorKey: 'name',
+      header: 'Task Name',
+      cell: ({ row }) => (
+        <div>
+          <div className="task-name">{row.original.name}</div>
+          <div className="task-meta">
+            ID: {row.original.id.slice(0, 8)}...
+          </div>
+        </div>
+      ),
+      size: 250,
+    },
+    {
+      accessorKey: 'description',
+      header: 'Description',
+      cell: ({ getValue }) => (
+        <div className="task-description">
+          {getValue()?.slice(0, 150)}
+          {getValue()?.length > 150 ? '...' : ''}
+        </div>
+      ),
+      size: 300,
+    },
+    {
+      accessorKey: 'status',
+      header: 'Status',
+      cell: ({ getValue }) => (
+        <span className={`status-badge status-${getValue()}`}>
+          {getValue()?.replace('_', ' ')}
+        </span>
+      ),
+      size: 120,
+    },
+    {
+      accessorKey: 'createdAt',
+      header: 'Created',
+      cell: ({ getValue }) => {
+        const date = new Date(getValue());
+        return (
+          <div className="task-meta">
+            {date.toLocaleDateString()}<br />
+            {date.toLocaleTimeString()}
+          </div>
+        );
+      },
+      size: 120,
+    },
+    {
+      accessorKey: 'updatedAt',
+      header: 'Updated',
+      cell: ({ getValue }) => {
+        const date = new Date(getValue());
+        return (
+          <div className="task-meta">
+            {date.toLocaleDateString()}<br />
+            {date.toLocaleTimeString()}
+          </div>
+        );
+      },
+      size: 120,
+    },
+    {
+      accessorKey: 'notes',
+      header: 'Notes',
+      cell: ({ getValue }) => {
+        const notes = getValue();
+        if (!notes) return <span style={{ color: '#666' }}>—</span>;
+        return (
+          <div className="task-description">
+            {notes.slice(0, 100)}
+            {notes.length > 100 ? '...' : ''}
+          </div>
+        );
+      },
+      size: 200,
+    },
+  ], []);
+
+  const table = useReactTable({
+    data,
+    columns,
+    state: {
+      globalFilter,
+    },
+    onGlobalFilterChange,
+    getCoreRowModel: getCoreRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    globalFilterFn: 'includesString',
+    initialState: {
+      pagination: {
+        pageSize: 20,
+      },
+    },
+  });
+
+  if (data.length === 0) {
+    return (
+      <div className="loading">
+        No tasks found in this profile
+      </div>
+    );
+  }
+
+  return (
+    <>
+      <table className="table">
+        <thead>
+          {table.getHeaderGroups().map(headerGroup => (
+            <tr key={headerGroup.id}>
+              {headerGroup.headers.map(header => (
+                <th 
+                  key={header.id}
+                  style={{ width: header.getSize() }}
+                  onClick={header.column.getCanSort() ? header.column.getToggleSortingHandler() : undefined}
+                  className={header.column.getCanSort() ? 'sortable' : ''}
+                >
+                  {flexRender(header.column.columnDef.header, header.getContext())}
+                  {header.column.getIsSorted() && (
+                    <span style={{ marginLeft: '8px' }}>
+                      {header.column.getIsSorted() === 'desc' ? '↓' : '↑'}
+                    </span>
+                  )}
+                </th>
+              ))}
+            </tr>
+          ))}
+        </thead>
+        <tbody>
+          {table.getRowModel().rows.map(row => (
+            <tr key={row.id}>
+              {row.getVisibleCells().map(cell => (
+                <td key={cell.id}>
+                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                </td>
+              ))}
+            </tr>
+          ))}
+        </tbody>
+      </table>
+
+      <div className="pagination">
+        <div className="pagination-info">
+          Showing {table.getState().pagination.pageIndex * table.getState().pagination.pageSize + 1} to{' '}
+          {Math.min(
+            (table.getState().pagination.pageIndex + 1) * table.getState().pagination.pageSize,
+            table.getFilteredRowModel().rows.length
+          )}{' '}
+          of {table.getFilteredRowModel().rows.length} tasks
+          {globalFilter && ` (filtered from ${data.length} total)`}
+        </div>
+        
+        <div className="pagination-controls">
+          <button
+            onClick={() => table.setPageIndex(0)}
+            disabled={!table.getCanPreviousPage()}
+          >
+            {'<<'}
+          </button>
+          <button
+            onClick={() => table.previousPage()}
+            disabled={!table.getCanPreviousPage()}
+          >
+            {'<'}
+          </button>
+          <span>
+            Page {table.getState().pagination.pageIndex + 1} of{' '}
+            {table.getPageCount()}
+          </span>
+          <button
+            onClick={() => table.nextPage()}
+            disabled={!table.getCanNextPage()}
+          >
+            {'>'}
+          </button>
+          <button
+            onClick={() => table.setPageIndex(table.getPageCount() - 1)}
+            disabled={!table.getCanNextPage()}
+          >
+            {'>>'}
+          </button>
+        </div>
+      </div>
+    </>
+  );
+}
+
+export default TaskTable;
